Adds helper functions for remaining orion bms tests

+++ src/Sensor/CanSensorOrionBms.cpp
@@ -129,7 +129,7 @@ void CanSensorOrionBms::update(CanMessage message) {
 }
 
 int16_t CanSensorOrionBms::_parseInt16(uint8_t* buf) {
-	return (float)( *buf << 8 | *(buf + 1) );
+	return (int16_t)( *buf << 8 | *(buf + 1) );
 }
 
 // Faults are prioritized from left to right, most significant bit to least significant
